+++ robotidy/cli.py
@@ -60,7 +60,7 @@ click.rich_click.OPTION_GROUPS = {
                 "--endline",
             ],
         },
-        {"name": "File exclusion", "options": ["--exclude", "--extend-exclude"]},
+        {"name": "File exclusion", "options": ["--exclude", "--extend-exclude", "--skip-gitignore"]},
         {
             "name": "Other",
             "options": ["--target-version", "--verbose", "--color", "--output", "--version", "--help"],
@@ -244,6 +244,12 @@ def print_transformers_list(target_version: int):
     ),
 )
 @click.option(
+    "--skip-gitignore",
+    is_flag=True,
+    show_default=True,
+    help="Skip **.gitignore** files and do not ignore files listed inside.",
+)
+@click.option(
     "--config",
     type=click.Path(
         exists=True,
@@ -390,6 +396,7 @@ def cli(
     src: Tuple[str, ...],
     exclude: Optional[Pattern],
     extend_exclude: Optional[Pattern],
+    skip_gitignore: bool,
     overwrite: bool,
     diff: bool,
     color: bool,
@@ -457,6 +464,7 @@ def cli(
         src=src,
         exclude=exclude,
         extend_exclude=extend_exclude,
+        skip_gitignore=skip_gitignore,
         overwrite=overwrite,
         show_diff=diff,
         formatting_config=formatting_config,