+++ tests/utest/test_utils.py
@@ -23,6 +23,7 @@ def app():
         src=(".",),
         exclude=None,
         extend_exclude=None,
+        skip_gitignore=False,
         overwrite=False,
         show_diff=False,
         formatting_config=formatting_config,
@@ -102,6 +103,12 @@ class TestUtils:
     def test_get_line_ending(self, line_sep, source_file, expected, app):
         source = str(Path(__file__).parent / "testdata" / "auto_line_sep" / source_file)
         app.formatting_config = GlobalFormattingConfig(
-            space_count=4, indent=4, line_sep=line_sep, start_line=None, separator="space", end_line=None, line_length=120
+            space_count=4,
+            indent=4,
+            line_sep=line_sep,
+            start_line=None,
+            separator="space",
+            end_line=None,
+            line_length=120,
         )
         assert app.get_line_ending(source) == expected